[WIP] improving code
